Rename quarkus-qute-server-pages to quarkus-qute-web
- related to https://github.com/quarkusio/quarkus/issues/35739
- rename the groupId from io.quarkiverse.quteserverpages to
io.quarkiverse.qute.web
- rename the artifactId from quarkus-qute-server-pages to
quarkus-qute-web
- rename the package from io.quarkiverse.quteserverpages to
io.quarkiverse.qute.web
- rename the config prefix from quarkus.qsp to quarkus.qute-web
- rename class names with "Qsp" prefix to use the prefix "QuteWeb"
- convert config to ConfigMapping
- introduce new config property: web-templates-dir; defaulted to "web"
- changed the default value for root-path to "/"
- changed the default value for hidden-templates; no templates are
hidden
- rename the "qsp:" namespace to "http:"
