+++ README.md
@@ -1,9 +1,9 @@
-# Quarkus QuteServerPages
+# Quarkus Qute Web
 
 [![Version](https://img.shields.io/maven-central/v/io.quarkiverse.quteserverpages/quarkus-qute-server-pages.svg?label=Maven%20Central)](https://search.maven.org/artifact/io.quarkiverse.quteserverpages/quarkus-qute-server-pages)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 
-The goal of this simple extension is to expose [Qute](https://quarkus.io/guides/qute-reference) templates located in the `src/main/resource/templates` directory via HTTP. Automatically, no controllers needed. For example, the template `src/main/resource/templates/foo.html` will be served from the paths `/qsp/foo` and `/qsp/foo.html` by default.
+The goal of this extension is to expose the [Qute](https://quarkus.io/guides/qute-reference) templates located in the `src/main/resource/templates` directory via HTTP. Automatically, no controllers needed. For example, the template `src/main/resource/templates/web/foo.html` will be served from the paths `/foo` and `/foo.html` by default.
 
 In a template you can access:
 
@@ -13,6 +13,6 @@ In a template you can access:
 - [Namespace Extension Methods](https://quarkus.io/guides/qute-reference#namespace_extension_methods) in general
 - [global variables](https://quarkus.io/guides/qute-reference#global_variables)
 - the current `io.vertx.core.http.HttpServerRequest` via the `qsp:` namespace, e.g. `{qsp:request.path}`
-- the query parameters via the qsp: namespace, e.g. `{qsp:param('name')}` and `{qsp:param('name','DefaultName'}`
+- the query parameters via the `http:` namespace, e.g. `{http:param('name')}` and `{http:param('name','DefaultName'}`
 
 Read the full [documentation](https://quarkiverse.github.io/quarkiverse-docs/quarkus-quteserverpages/dev/index.html).
