+++ deployment/src/test/java/io/quarkiverse/qute/web/test/BlockingHandlerTest.java
@@ -1,0 +1,49 @@
+package io.quarkiverse.qute.web.test;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.Matchers.containsString;
+
+import jakarta.inject.Named;
+import jakarta.inject.Singleton;
+
+import org.jboss.shrinkwrap.api.asset.StringAsset;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.RegisterExtension;
+
+import io.quarkus.test.QuarkusUnitTest;
+import io.vertx.core.Context;
+
+public class BlockingHandlerTest {
+
+    @RegisterExtension
+    static final QuarkusUnitTest config = new QuarkusUnitTest().withApplicationRoot(root -> {
+        root
+                .addAsResource(new StringAsset(
+                        "blocking={cdi:bean.isOnWorkerThread}"),
+                        "templates/web/blocking.txt")
+                .addAsResource(new StringAsset(
+                        "quarkus.qute.web.use-blocking-handler=true"),
+                        "application.properties");
+    });
+
+    @Test
+    public void testHandler() {
+        given()
+                .when().get("/blocking")
+                .then()
+                .statusCode(200)
+                .body(containsString("blocking=true"));
+
+    }
+
+    @Named
+    @Singleton
+    public static class Bean {
+
+        public boolean isOnWorkerThread() {
+            return Context.isOnWorkerThread();
+        }
+
+    }
+
+}
