+++ docs/pom.xml
@@ -2,20 +2,20 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.quarkiverse.quteserverpages</groupId>
-        <artifactId>quarkus-qute-server-pages-parent</artifactId>
+        <groupId>io.quarkiverse.qute.web</groupId>
+        <artifactId>quarkus-qute-web-parent</artifactId>
         <version>999-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>quarkus-qute-server-pages-docs</artifactId>
-    <name>Quarkus QuteServerPages - Documentation</name>
+    <artifactId>quarkus-qute-web-docs</artifactId>
+    <name>Quarkus Qute Web - Documentation</name>
 
     <dependencies>
         <!-- Make sure the doc is built after the other artifacts -->
         <dependency>
-            <groupId>io.quarkiverse.quteserverpages</groupId>
-            <artifactId>quarkus-qute-server-pages-deployment</artifactId>
+            <groupId>io.quarkiverse.qute.web</groupId>
+            <artifactId>quarkus-qute-web-deployment</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
