+++ integration-tests/pom.xml
@@ -2,12 +2,12 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>io.quarkiverse.quteserverpages</groupId>
-    <artifactId>quarkus-qute-server-pages-parent</artifactId>
+    <groupId>io.quarkiverse.qute.web</groupId>
+    <artifactId>quarkus-qute-web-parent</artifactId>
     <version>999-SNAPSHOT</version>
   </parent>
-  <artifactId>quarkus-qute-server-pages-integration-tests</artifactId>
-  <name>Quarkus QuteServerPages - Integration Tests</name>
+  <artifactId>quarkus-qute-web-integration-tests</artifactId>
+  <name>Quarkus Qute Web - Integration Tests</name>
   <properties>
     <skipITs>true</skipITs>
   </properties>
@@ -17,8 +17,8 @@
       <artifactId>quarkus-resteasy</artifactId>
     </dependency>
     <dependency>
-      <groupId>io.quarkiverse.quteserverpages</groupId>
-      <artifactId>quarkus-qute-server-pages</artifactId>
+      <groupId>io.quarkiverse.qute.web</groupId>
+      <artifactId>quarkus-qute-web</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
