+++ integration-tests/src/main/resources/templates/web/colors.html
