+++ integration-tests/src/main/resources/templates/web/integers.html
