+++ integration-tests/src/main/resources/templates/web/nested/enum.txt
