+++ integration-tests/src/test/java/io/quarkiverse/quteserverpages/it/QspTest.java
@@ -14,58 +14,58 @@ public class QspTest {
     @Test
     public void testTemplates() {
         given()
-                .when().get("/qsp/hello")
+                .when().get("/hello")
                 .then()
                 .statusCode(200)
                 .body(containsString("Hello world!"));
 
         // CDI bean
         given()
-                .when().get("/qsp/names.html")
+                .when().get("/names.html")
                 .then()
                 .statusCode(200)
                 .body(containsString("<li>Violet</li>"));
 
         // Globals
         given()
-                .when().get("/qsp/integers?name=foo")
+                .when().get("/integers?name=foo")
                 .then()
                 .statusCode(200)
                 .body(containsString("1:11:42"), containsString("Name: foo"));
 
         // Static method
         given()
-                .when().get("/qsp/colors")
+                .when().get("/colors")
                 .then()
                 .statusCode(200)
                 .body(containsString("red, green, blue"));
 
         // Enum
         given()
-                .when().get("/qsp/nested/enum")
+                .when().get("/nested/enum")
                 .then()
                 .statusCode(200)
                 .body(is("OK"));
 
         // Not found
         given()
-                .when().get("/qsp/fooooooo")
+                .when().get("/fooooooo")
                 .then()
                 .statusCode(404);
 
         // Fragment
         given()
-                .when().get("/qsp/fragment")
+                .when().get("/fragment")
                 .then()
                 .statusCode(200)
                 .body(containsString("<title>Fragment</title> "), containsString("<p>Foo!</p>"));
         given()
-                .when().get("/qsp/fragment?frag=foo")
+                .when().get("/fragment?frag=foo")
                 .then()
                 .statusCode(200)
                 .body(is("<p>Foo!</p>"));
         given()
-                .when().get("/qsp/fragment?frag=bar")
+                .when().get("/fragment?frag=bar")
                 .then()
                 .statusCode(404);
     }
