+++ pom.xml
@@ -6,11 +6,11 @@
     <artifactId>quarkiverse-parent</artifactId>
     <version>15</version>
   </parent>
-  <groupId>io.quarkiverse.quteserverpages</groupId>
-  <artifactId>quarkus-qute-server-pages-parent</artifactId>
+  <groupId>io.quarkiverse.qute.web</groupId>
+  <artifactId>quarkus-qute-web-parent</artifactId>
   <version>999-SNAPSHOT</version>
   <packaging>pom</packaging>
-  <name>Quarkus QuteServerPages - Parent</name>
+  <name>Quarkus Qute Web - Parent</name>
   <modules>
     <module>deployment</module>
     <module>runtime</module>
